sim:epiphany: Use 64-bit address word size
This adds needed basic support for 64-bit addressing mode to Epiphany
simulator port. The Epiphany-V chip has a 64-bit addressing mode that is
controlled by toggling a bit in the config register. For simplicity the plan
is to use 64-bit addressing internally for Epiphany-III and Epiphany-IV too,
but to truncate addresses to 32-bits in the cgen cpu defintion file for older
chips.

sim/epiphany/
	* configure.ac (SIM_AC_OPTION_ALIGNMENT): Set 64-bit address word
	bitsize.
	* sim-main.h (CIA_ADDR): Define.
	(INVALID_INSTRUCTION_ADDRESS): Define.
	* dv-epiphany_mem.c (epiphany_mem_finish): Fix compiler warning.
	(epiphany_mem_signal): Change to address_word type for addr
	and nr_bytes.
	(epiphany_mem_io_read_buffer): Change to address word type for
	addr, nr_bytes, and return value.
	(epiphany_mem_io_write_buffer): Likewise.
	* epiphany.c (epiphanybf_scache_invalidate): Use
	INVALID_INSTRUCTION_ADDRESS instead of hardcoded value.
	* esim/esim.h (es_cluster_cfg): Use uint64_t type for
	addresses and sizes in the public API.
	(es_mem_store): Likewise.
	(es_mem_load): Likewise.
	(es_mem_testset): Likewise.
	* esim/esim-int.h (es_transl): Use address_word type for addresses and
	sizes internally.
	* esim/esim-net.h (es_net_addr_translate): Likewise.
	* esim/esim.c (es_tx_one_shm_store): Likewise.
	(es_addr_to_node(const es_state *esim, address_word addr): Likewise.
	(es_mem_store): Likewise.
	(es_mem_load): Likewise.
	(es_mem_testset): Likewise.
	(es_dump_config):
	* esim/esim-net.c (es_net_addr_translate): Adjust function header for
	above.

+++ sim/epiphany/esim/esim.h
@@ -35,8 +35,8 @@ typedef struct es_cluster_cfg_ {
     size_t   core_mem_region;         /*!< Core memory region size         */
     /* signed   core_phys_mem; */     /* Allocate entire region for now    */
     unsigned ext_ram_node;            /*!< Let this be rank '0' for now    */
-    uint32_t ext_ram_base;            /*!< core_mem_region must be divisor */
-    size_t   ext_ram_size;            /*!< Size of external memory         */
+    uint64_t ext_ram_base;            /*!< core_mem_region must be divisor */
+    uint64_t ext_ram_size;            /*!< Size of external memory         */
 
     /*! @privatesection */
     /* Keep your grubby little mitts off of these plz :) */
@@ -66,9 +66,12 @@ void es_fini(es_state *esim);
 int es_client_connect(es_state **esim);
 void es_client_disconnect(es_state *esim);
 
-int es_mem_store(es_state *esim, uint32_t addr, uint32_t size, uint8_t *src);
-int es_mem_load(es_state *esim, uint32_t addr, uint32_t size, uint8_t *dst);
-int es_mem_testset(es_state *esim, uint32_t addr, uint32_t size, uint8_t *dst);
+int es_mem_store(es_state *esim, uint64_t addr, uint64_t size,
+		 uint8_t *src);
+int es_mem_load(es_state *esim, uint64_t addr, uint64_t size,
+		uint8_t *dst);
+int es_mem_testset(es_state *esim, uint64_t addr, uint64_t size,
+		   uint8_t *dst);
 
 void es_wait_run(es_state *esim);
 void es_wait_exit(es_state *esim);
